pyproject.toml
[project]
name = "phone-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.66.0",
    "kivy>=2.3.1",
    "pyaudio>=0.2.14",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.2",
    "websockets>=16.0",
]

[tool.uv]
package = true

[project.scripts]
phone-assistant = "phone_assistant.main:entry_point"
